updated PS kernels 3

+++ src/runtime_src/xdp/profile/plugin/aie_profile/x86/aie_profile_kernel_config.h
@@ -65,7 +65,7 @@ namespace xdp {
     struct ProfileTileType {
       uint16_t row;
       uint16_t col;
-      std::vector<uint16_t> stream_ids;
+      std::vector<uint8_t> stream_ids;
       uint16_t is_master;
       uint64_t itr_mem_addr;
       bool is_trigger;
